+++ library/c++/memshare/test_a.c
@@ -14,10 +14,10 @@ typedef struct {
 	char s[128];
 } Info;
 
-void msg_callback(const char* proc, int msg_type, int msg_len, const void* data) {
+void msg_callback(const char* proc, int msg_type, long msg_len, const void* data) {
 	Info* info;
 	int retvalue = 0;
-	printf("receive [%s], msg_type => %d, msg_len => %d \n", proc, msg_type, msg_len);
+	printf("receive [%s], msg_type => %d, msg_len => %ld \n", proc, msg_type, msg_len);
 	if (1 == msg_type) {
 		info = (Info*)data;
 		printf("receive data => %d, %ld, %f, %f, %s \n", info->a, info->b, info->c, info->d, info->s);
@@ -41,7 +41,7 @@ int main(int argc, char *argv[]) {
 
 	while (1) {
 		sleep(1);
-		send_msg("test_b", 3, 0, NULL);
-		send_msg("test_c", 4, 0, NULL);
+		send_msg("test_b", 3, 0L, NULL);
+		send_msg("test_c", 4, 0L, NULL);
 	}
 }
