+++ library/c++/memshare/test_b.c
@@ -13,9 +13,9 @@ typedef struct {
 	char s[64];
 } Info;
 
-void msg_callback(const char* proc, int msg_type, int msg_len, const void* data) {
+void msg_callback(const char* proc, int msg_type, long msg_len, const void* data) {
 	Info* info;
-	printf("receive [%s], msg_type => %d, msg_len => %d \n", proc, msg_type, msg_len);
+	printf("receive [%s], msg_type => %d, msg_len => %ld \n", proc, msg_type, msg_len);
 	if (7 == msg_type) {
 		info = (Info*)data;
 		printf("receive data => %d, %ld, %f, %f, %s \n", info->a, info->b, info->c, info->d, info->s);
@@ -44,7 +44,7 @@ int main(int argc, char *argv[]) {
 		memcpy(i->s, m, strlen(m) + 1);
 		send_msg("test_a", 1, sizeof(*i), (void*)i);
 		free(i);
-		send_msg("test_c", 2, 0, NULL);
+		send_msg("test_c", 2, 0L, NULL);
 	}
 
 	return 0;
